requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
PyGithub
PyYAML
requests
streamlit
